Barcode: improve consistency by moving logic into a BarcodeRepository struct
